pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "mediapipe>=0.10.21",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
]
